Added Special Attack with multiplier to Special Abilities. Uses visitor, state and command patterns.

+++ src/main/java/boardgame/gameModel/pieces/Archer.java
@@ -30,7 +30,7 @@ public class Archer extends Human {
     }
 
     @Override
-    public int getAttack() {
+    public double getAttack() {
         return attack;
     }
