+++ src/main/java/boardgame/gameModel/pieces/Medusa.java
@@ -41,7 +41,7 @@ public class Medusa extends Monster {
     }
 
     @Override
-    public int getAttack() {
+    public double getAttack() {
         return attack;
     }
